+++ src/sprints/sprints.service.ts
@@ -157,6 +157,7 @@ export class SprintsService {
                         title: true,
                         description: true,
                         status: true,
+                        updatedAt: true,
                     },
                 },
                 formResponseMeeting: {